Fix 'crash' when machine info would fail (e.g. corrupt XML)
Do not attempt to test a machine for which we get an error on doing an
openmsx_info machines <machine name>.

A machine was created even if it could (and would) never load a config
in it (because it is apparently corrupt). Creating the machine also
means the test will be done, which would (in the end) trigger _infoDone. But
when failing to get the machine info, _infoDone would also get triggered
more directly, causing a double decrease of _repliesLeft.

So, postpone creating a machine to test until we know the info topic
works to avoid this trouble. (Perhaps there's an easier way too, but
this seems to be just too.)