Fix the selection modes of the extension lists
The add extension selection list was not prepared to handle more than one selected item at a time. Even if it was, there's still the problem that an extension appears pre-selected at startup, which is a grave usability problem, so unless that is fixed, we restrict the list to one single selection at a time.

On the other hand, the active extensions list was prepared to deal with multiple selections, but the property allowing multiselect was not set.